fix: change timeout parameter type in _request_json method from int to float for better precision

+++ slidegen/services/factories/base_factory.py
@@ -7,7 +7,7 @@ class BaseFactory:
     """Shared helpers for provider-backed factories."""
 
     @staticmethod
-    async def _request_json(url: str, headers: dict[str, str], timeout: int = 20.0) -> dict[str, Any]:
+    async def _request_json(url: str, headers: dict[str, str], timeout: float = 20.0) -> dict[str, Any]:
         try:
             async with httpx.AsyncClient(timeout=timeout) as client:
                 response = await client.get(url, headers=headers)
